review: split up tests
